backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pymongo==4.6.0
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
email-validator==2.0.0.post2
pydantic==2.5.2
pydantic-settings==2.1.0
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dateutil==2.8.2
smtplib==0.0.1